Move setupSnapshotWithFreshState
