+++ java/core/src/test/java/sleeper/core/statestore/transactionlog/InMemoryTransactionLogSnapshotSetup.java
@@ -15,13 +15,38 @@
  */
 package sleeper.core.statestore.transactionlog;
 
+import sleeper.core.schema.Schema;
+import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.table.TableStatus;
 
+import static sleeper.core.statestore.FileReferenceTestData.DEFAULT_UPDATE_TIME;
+
 public class InMemoryTransactionLogSnapshotSetup {
+
     private final TableStatus sleeperTable;
     private final TransactionLogStore filesLog;
     private final TransactionLogStore partitionsLog;
+
+    public static InMemoryTransactionLogSnapshotSetup setupSnapshotWithFreshState(
+            TableStatus sleeperTable, Schema schema, SetupStateStore setupState) throws StateStoreException {
+        InMemoryTransactionLogStore fileTransactions = new InMemoryTransactionLogStore();
+        InMemoryTransactionLogStore partitionTransactions = new InMemoryTransactionLogStore();
+        StateStore stateStore = TransactionLogStateStore.builder()
+                .sleeperTable(sleeperTable)
+                .schema(schema)
+                .filesLogStore(fileTransactions)
+                .partitionsLogStore(partitionTransactions)
+                .build();
+        stateStore.fixFileUpdateTime(DEFAULT_UPDATE_TIME);
+        stateStore.fixPartitionUpdateTime(DEFAULT_UPDATE_TIME);
+        setupState.run(stateStore);
+        return new InMemoryTransactionLogSnapshotSetup(sleeperTable, fileTransactions, partitionTransactions);
+    }
+
+    public interface SetupStateStore {
+        void run(StateStore stateStore) throws StateStoreException;
+    }
 
     InMemoryTransactionLogSnapshotSetup(TableStatus sleeperTable, TransactionLogStore filesLog, TransactionLogStore partitionsLog) {
         this.sleeperTable = sleeperTable;
