+++ java/core/src/test/java/sleeper/core/statestore/transactionlog/InMemoryTransactionLogSnapshots.java
@@ -15,14 +15,7 @@
  */
 package sleeper.core.statestore.transactionlog;
 
-import sleeper.core.schema.Schema;
-import sleeper.core.statestore.StateStore;
-import sleeper.core.statestore.StateStoreException;
-import sleeper.core.table.TableStatus;
-
 import java.util.Optional;
-
-import static sleeper.core.statestore.FileReferenceTestData.DEFAULT_UPDATE_TIME;
 
 /**
  * An in-memory implementation of snapshots derived from a transaction log.
@@ -30,22 +23,6 @@ import static sleeper.core.statestore.FileReferenceTestData.DEFAULT_UPDATE_TIME;
 public class InMemoryTransactionLogSnapshots implements TransactionLogSnapshotLoader {
 
     private TransactionLogSnapshot latestSnapshot;
-
-    public static InMemoryTransactionLogSnapshotSetup setupSnapshotWithFreshState(
-            TableStatus sleeperTable, Schema schema, SetupStateStore setupState) throws StateStoreException {
-        InMemoryTransactionLogStore fileTransactions = new InMemoryTransactionLogStore();
-        InMemoryTransactionLogStore partitionTransactions = new InMemoryTransactionLogStore();
-        StateStore stateStore = TransactionLogStateStore.builder()
-                .sleeperTable(sleeperTable)
-                .schema(schema)
-                .filesLogStore(fileTransactions)
-                .partitionsLogStore(partitionTransactions)
-                .build();
-        stateStore.fixFileUpdateTime(DEFAULT_UPDATE_TIME);
-        stateStore.fixPartitionUpdateTime(DEFAULT_UPDATE_TIME);
-        setupState.run(stateStore);
-        return new InMemoryTransactionLogSnapshotSetup(sleeperTable, fileTransactions, partitionTransactions);
-    }
 
     public void setLatestSnapshot(TransactionLogSnapshot latestSnapshot) {
         this.latestSnapshot = latestSnapshot;
@@ -56,9 +33,4 @@ public class InMemoryTransactionLogSnapshots implements TransactionLogSnapshotLo
         return Optional.ofNullable(latestSnapshot)
                 .filter(snapshot -> snapshot.getTransactionNumber() >= transactionNumber);
     }
-
-    public interface SetupStateStore {
-        void run(StateStore stateStore) throws StateStoreException;
-    }
-
 }
